p1-ftds030-rmt-m2-RizkyChesterAbadi/deployment/requirements.txt
streamlit
pandas
seaborn==0.13.2
matplotlib
numpy
scikit-learn==1.4.1
Pillow
plotly
